+++ stopify-continuations/src/runtime/runtime.ts
@@ -9,25 +9,18 @@ import { EagerRuntime } from './eagerRuntime';
 import { RetvalRuntime } from './retvalRuntime';
 import { FudgeRuntime } from './fudgeRuntime';
 
+export * from './abstractRuntime';
+
 let savedRTS: Runtime | undefined;
-export function newRTS(transform: string) : Runtime {
-  if (savedRTS) {
-    return savedRTS;
-  }
-  if (transform === 'lazy') {
-    savedRTS = new LazyRuntime();
-  }
-  else if (transform === 'eager') {
-    savedRTS = new EagerRuntime();
-  }
-  else if (transform === 'retval') {
-    savedRTS = new RetvalRuntime();
-  }
-  else if (transform === 'fudge') {
-    savedRTS = new FudgeRuntime();
-  }
-  else {
-    throw new Error(`bad runtime: ${transform}`);
+export function newRTS(transform: string): Runtime {
+  if (!savedRTS) {
+    switch (transform) {
+      case 'lazy': savedRTS = new LazyRuntime(); break;
+      case 'eager': savedRTS = new EagerRuntime(); break;
+      case 'retval': savedRTS = new RetvalRuntime(); break;
+      case 'fudge': savedRTS = new FudgeRuntime(); break;
+      default: throw new Error(`bad runtime: ${transform}`);
+    }
   }
   return savedRTS;
 }
